Chnage the report structure
