+++ static_report/src/components/Columns/MasterSideNav/ColumnListAccordionPanel.tsx
@@ -1,44 +1,34 @@
 import { AccordionPanel, Box } from '@chakra-ui/react';
-import { Comparable, Selectable } from '../../../types';
+import { useRoute } from 'wouter';
+import { Comparable } from '../../../types';
 import { CompColEntryItem } from '../../../utils';
+import { COLUMN_DETAILS_ROUTE_PATH } from '../../../utils/routes';
 import { ColumnDetailListItem } from './ColumnDetailListItem';
 
 /**
  * ColumnItemList: Accordion UI Child Body
  */
-interface ColumnListAccordionPanelProps extends Comparable, Selectable {
+interface ColumnListAccordionPanelProps extends Comparable {
+  tableName: string;
   compColList?: CompColEntryItem[];
-  currentColumn?: string;
-  currentTable?: string;
   indexedTableName: string;
 }
 export function ColumnListAccordionPanel({
+  tableName,
   compColList = [],
-  onSelect,
   singleOnly,
-  currentColumn,
-  currentTable,
-  indexedTableName,
 }: ColumnListAccordionPanelProps) {
   return (
     <AccordionPanel>
       <Box>
         {compColList.map(([colKey, { base, target }]) => {
-          const isActiveColumn =
-            (target || base)?.name === currentColumn &&
-            indexedTableName === currentTable;
-
           return (
             <Box key={colKey}>
               {/* LIST - Columns */}
               <ColumnDetailListItem
-                isActive={isActiveColumn}
-                tableName={currentTable}
+                tableName={tableName}
                 baseColumnDatum={base}
                 targetColumnDatum={target}
-                onSelect={(data) => {
-                  onSelect({ ...data, tableName: indexedTableName });
-                }}
                 singleOnly={singleOnly}
                 p={3}
               />
