+++ static_report/src/components/Common/Main.tsx
@@ -4,6 +4,7 @@ import * as amplitude from '@amplitude/analytics-browser';
 
 import { Navbar } from './Navbar';
 import { mainContentAreaHeight } from '../../utils';
+import { useLocation } from 'wouter';
 
 interface Props extends FlexProps {
   children: ReactNode;
