+++ static_report/src/pages/SRTablesListPage.tsx
@@ -13,7 +13,6 @@ import { TableColumnSchemaList } from '../components/Tables/TableList/TableColum
 import { useAmplitudeOnMount } from '../hooks/useAmplitudeOnMount';
 import { AMPLITUDE_EVENTS, SR_TYPE_LABEL } from '../utils/amplitudeEvents';
 import { CommonModal } from '../components/Common/CommonModal';
-import { MasterDetailContainer } from '../components/Common/MasterDetailContainer';
 
 type Props = { data: SaferSRSchema };
 
@@ -22,13 +21,7 @@ export function SRTablesListPage({ data }: Props) {
   const [tableColsEntryId, setTableColsEntryId] = useState(-1);
   const setReportData = useReportStore((s) => s.setReportRawData);
   setReportData({ base: data });
-  const {
-    tableColumnsOnly = [],
-    assertionsOnly,
-    rawData,
-  } = useReportStore.getState();
-
-  const [, setLocation] = useLocation();
+  const { tableColumnsOnly = [], assertionsOnly } = useReportStore.getState();
 
   useDocumentTitle('Single-Run Report: Tables');
   useAmplitudeOnMount({
@@ -40,68 +33,57 @@ export function SRTablesListPage({ data }: Props) {
   });
 
   return (
-    <Main isSingleReport>
-      <MasterDetailContainer
-        initAsExpandedTables
-        rawData={rawData}
-        tableColEntries={tableColumnsOnly}
-      >
-        <Flex direction="column" width={'100%'} minHeight="650px">
-          <Grid
-            templateColumns={tableListGridTempCols}
-            maxW={tableListMaxWidth}
-            px={4}
-            my={6}
-          >
-            <Text>Name</Text>
-            <Text>Summary</Text>
-            <Text>Assertions</Text>
-          </Grid>
-          {tableColumnsOnly.map((tableColsEntry, i) => {
-            return (
-              <TableListItem
-                key={i}
-                combinedAssertions={assertionsOnly}
-                combinedTableEntry={tableColsEntry}
-                singleOnly
-                onInfoClick={() => {
-                  setTableColsEntryId(i);
-                  modal.onOpen();
-                }}
-                onSelect={({ tableName }) =>
-                  setLocation(`/tables/${tableName}/columns/`)
-                }
-              />
-            );
-          })}
-        </Flex>
-
-        <CommonModal
-          {...modal}
-          size="2xl"
-          title={
-            tableColsEntryId !== -1 && tableColumnsOnly[tableColsEntryId][0]
-          }
-          onClose={() => {
-            setTableColsEntryId(-1);
-            modal.onClose();
-          }}
+    <>
+      <Flex direction="column" width={'100%'} minHeight="650px">
+        <Grid
+          templateColumns={tableListGridTempCols}
+          maxW={tableListMaxWidth}
+          px={4}
+          my={6}
         >
-          <Text fontSize="lg" mb={4}>
-            Description:{' '}
-            {(tableColsEntryId !== -1 &&
-              tableColumnsOnly[tableColsEntryId][1].base?.description) ?? (
-              <Text as="i">No description provided.</Text>
-            )}
-          </Text>
-          {tableColsEntryId !== -1 && (
-            <TableColumnSchemaList
+          <Text>Name</Text>
+          <Text>Summary</Text>
+          <Text>Assertions</Text>
+        </Grid>
+        {tableColumnsOnly.map((tableColsEntry, i) => {
+          return (
+            <TableListItem
+              key={i}
+              combinedAssertions={assertionsOnly}
+              combinedTableEntry={tableColsEntry}
               singleOnly
-              baseTableEntryDatum={tableColumnsOnly[tableColsEntryId][1].base}
+              onInfoClick={() => {
+                setTableColsEntryId(i);
+                modal.onOpen();
+              }}
             />
+          );
+        })}
+      </Flex>
+
+      <CommonModal
+        {...modal}
+        size="2xl"
+        title={tableColsEntryId !== -1 && tableColumnsOnly[tableColsEntryId][0]}
+        onClose={() => {
+          setTableColsEntryId(-1);
+          modal.onClose();
+        }}
+      >
+        <Text fontSize="lg" mb={4}>
+          Description:{' '}
+          {(tableColsEntryId !== -1 &&
+            tableColumnsOnly[tableColsEntryId][1].base?.description) ?? (
+            <Text as="i">No description provided.</Text>
           )}
-        </CommonModal>
-      </MasterDetailContainer>
-    </Main>
+        </Text>
+        {tableColsEntryId !== -1 && (
+          <TableColumnSchemaList
+            singleOnly
+            baseTableEntryDatum={tableColumnsOnly[tableColsEntryId][1].base}
+          />
+        )}
+      </CommonModal>
+    </>
   );
 }
