+++ static_report/src/utils/routes.ts
@@ -1,4 +1,6 @@
+export const TABLE_LIST_ROUTE_PATH = '/tables';
 export const TABLE_DETAILS_ROUTE_PATH = '/tables/:tableName';
-export const COLUMN_DETAILS_ROUTE_PATH = `${TABLE_DETAILS_ROUTE_PATH}/columns/:columnName*`;
+export const COLUMN_DETAILS_ROUTE_PATH =
+  '/tables/:tableName/columns/:columnName';
 export const ASSERTIONS_ROUTE_PATH = `/assertions`;
 export const BM_ROUTE_PATH = `/metrics`;
